fix error

+++ 0x00-hello_world/6-size.c
@@ -9,6 +9,6 @@ int main(void)
 	printf("Size of an int: %zu byte(s)\n", sizeof(int));
 	printf("Size of a long int: %zu byte(s)\n", sizeof(long int));
 	printf("Size of a long long int: %zu byte(s)\n", sizeof(long long int));
-	printf("Size of a float: %f byte(s)\n", sizeof(long float));
+	printf("Size of a float: %f byte(s)\n", sizeof(float));
 	return (0);
 }
